Create SourceVM base class
