+++ PlayCover/Views/PlayCoverApp.swift
@@ -35,7 +35,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
 struct PlayCoverApp: App {
     @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
     @StateObject var updaterViewModel = UpdaterViewModel()
-    var storeVM = StoreVM.shared
+    var ipaSourceVM = IPASourceVM.shared
 
     @State var xcodeCliInstalled = shell.isXcodeCliToolsInstalled
     @State var isSigningSetupShown = false
@@ -46,7 +46,7 @@ struct PlayCoverApp: App {
                      isSigningSetupShown: $isSigningSetupShown)
                 .environmentObject(InstallVM.shared)
                 .environmentObject(AppsVM.shared)
-                .environmentObject(storeVM)
+                .environmentObject(ipaSourceVM)
                 .environmentObject(AppIntegrity())
                 .onAppear {
                     NSWindow.allowsAutomaticWindowTabbing = false
@@ -64,7 +64,7 @@ struct PlayCoverApp: App {
 
         Settings {
             PlayCoverSettingsView(updaterViewModel: updaterViewModel)
-                .environmentObject(storeVM)
+                .environmentObject(ipaSourceVM)
         }
     }
 }
