+++ PlayCover/Views/Settings/PlayCoverSettingsView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 
 struct PlayCoverSettingsView: View {
     @ObservedObject var updaterViewModel: UpdaterViewModel
-    @EnvironmentObject var storeVM: StoreVM
+    @EnvironmentObject var ipaSourceVM: IPASourceVM
 
     private enum Tabs: Hashable {
         case updates, ipasource
@@ -28,7 +28,7 @@ struct PlayCoverSettingsView: View {
                     Label("preferences.tab.ipasource", systemImage: "list.bullet")
                 }
                 .tag(Tabs.ipasource)
-                .environmentObject(storeVM)
+                .environmentObject(ipaSourceVM)
         }
     }
 }
